+++ src/main/java/roomescape/reservation/dto/response/PaymentErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.dto.response;
 
 public record PaymentErrorResponse(
         String code,
