+++ src/main/java/roomescape/reservation/error/handler/PaymentExceptionHandler.java
@@ -1,10 +1,12 @@
-package roomescape.reservation.exception;
+package roomescape.reservation.error.handler;
 
 import org.springframework.core.annotation.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.reservation.ErrorResponse;
+import roomescape.reservation.dto.response.ErrorResponse;
+import roomescape.reservation.error.exception.PaymentClientException;
+import roomescape.reservation.error.exception.PaymentServerException;
 
 @RestControllerAdvice
 @Order(1)
